LoginLogout Module implemented in SkillsDetails Fragment
Applied loginLogout presenter in signup activity

Refactored Settings and Skills Activity
