+++ app/src/main/java/org/fossasia/susi/ai/login/contract/ILoginLogoutModulePresenter.kt
@@ -1,0 +1,5 @@
+package org.fossasia.susi.ai.login.contract
+
+interface ILoginLogoutModulePresenter {
+    fun loginLogout()
+}
