+++ app/src/main/java/org/fossasia/susi/ai/signup/SignUpActivity.kt
@@ -19,6 +19,8 @@ import org.fossasia.susi.ai.helper.AlertboxHelper
 import org.fossasia.susi.ai.helper.Constant
 import org.fossasia.susi.ai.helper.CredentialHelper
 import org.fossasia.susi.ai.login.ForgotPass
+import org.fossasia.susi.ai.login.LoginLogoutModulePresenter
+import org.fossasia.susi.ai.login.contract.ILoginLogoutModulePresenter
 import org.fossasia.susi.ai.signup.contract.ISignUpPresenter
 import org.fossasia.susi.ai.signup.contract.ISignUpView
 import org.fossasia.susi.ai.skills.SkillsActivity
@@ -33,6 +35,7 @@ import org.fossasia.susi.ai.skills.SkillsActivity
 class SignUpActivity : AppCompatActivity(), ISignUpView {
 
     private lateinit var signUpPresenter: ISignUpPresenter
+    private lateinit var loginLogoutModulePresenter: ILoginLogoutModulePresenter
     private lateinit var progressDialog: ProgressDialog
     private lateinit var forgotPasswordProgressDialog: Dialog
     private lateinit var builder: AlertDialog.Builder
@@ -78,6 +81,7 @@ class SignUpActivity : AppCompatActivity(), ISignUpView {
 
         signUpPresenter = SignUpPresenter(this)
         signUpPresenter.onAttach(this)
+        loginLogoutModulePresenter = LoginLogoutModulePresenter(this)
     }
 
     private fun addListeners() {
@@ -220,7 +224,7 @@ class SignUpActivity : AppCompatActivity(), ISignUpView {
 
     private fun signUpToLoginPage() {
         signUpToLogin.setOnClickListener {
-            signUpPresenter.loginLogout()
+            loginLogoutModulePresenter.loginLogout()
             val intent = Intent(this@SignUpActivity, LoginActivity::class.java)
             intent.putExtra("email", email.editText?.text.toString())
             intent.flags = Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_NEW_TASK
