+++ app/src/main/java/org/fossasia/susi/ai/skills/skilldetails/SkillDetailsFragment.kt
@@ -1,6 +1,7 @@
 package org.fossasia.susi.ai.skills.skilldetails
 
 import android.annotation.SuppressLint
+import android.app.Activity
 import android.content.Intent
 import android.graphics.Color
 import android.net.Uri
@@ -34,6 +35,8 @@ import org.fossasia.susi.ai.dataclasses.ReportSkillQuery
 import org.fossasia.susi.ai.helper.PrefManager
 import org.fossasia.susi.ai.helper.Utils
 import org.fossasia.susi.ai.login.LoginActivity
+import org.fossasia.susi.ai.login.LoginLogoutModulePresenter
+import org.fossasia.susi.ai.login.contract.ILoginLogoutModulePresenter
 import org.fossasia.susi.ai.rest.responses.susi.GetSkillFeedbackResponse
 import org.fossasia.susi.ai.rest.responses.susi.SkillData
 import org.fossasia.susi.ai.rest.responses.susi.Stars
@@ -45,6 +48,7 @@ import org.fossasia.susi.ai.skills.skilldetails.contract.ISkillDetailsView
 class SkillDetailsFragment : Fragment(), ISkillDetailsView {
 
     private lateinit var skillDetailsPresenter: ISkillDetailsPresenter
+    private lateinit var loginLogoutModulePresenter: ILoginLogoutModulePresenter
 
     private lateinit var skillData: SkillData
     private lateinit var skillGroup: String
@@ -72,6 +76,7 @@ class SkillDetailsFragment : Fragment(), ISkillDetailsView {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
         skillDetailsPresenter = SkillDetailsPresenter(this)
+        loginLogoutModulePresenter = LoginLogoutModulePresenter(Activity())
         skillDetailsPresenter.onAttach(this)
         skillData = arguments?.getParcelable(
                 SKILL_KEY) as SkillData
@@ -499,7 +504,7 @@ class SkillDetailsFragment : Fragment(), ISkillDetailsView {
         } else {
             tvAnonymousPostFeedback.visibility = View.VISIBLE
             tvAnonymousPostFeedback.setOnClickListener {
-                skillDetailsPresenter.loginLogout()
+                loginLogoutModulePresenter.loginLogout()
                 val intentToLogin = Intent(context, LoginActivity::class.java)
                 intentToLogin.flags = Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_NEW_TASK
                 startActivity(intentToLogin)
