+++ app/src/main/java/org/fossasia/susi/ai/skills/skilldetails/contract/ISkillDetailsPresenter.kt
@@ -33,5 +33,4 @@ interface ISkillDetailsPresenter {
 
     fun onDetach()
 
-    fun loginLogout()
 }
